requirements.txt
Django==2.1.2
pytz==2018.5
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-expiring-authtoken==0.1.4
mysql-connector-python==8.0.12
